Using word indices to identify word pairs
Previously, the server kept track of the words practiced in the client and the
client only sent back something like "the user said YES to the second word you
sent me". This could have caused problems if a new client reinitialized the
words to be practiced in the server, because the first client would have still
used its original indexing.

Now the words are identified by the index of the word pair in the word list
object and the direction of the question, which are both send back to the
server by the JavaScript client when the user answers a question.

The model of handling the word list file has also changed: now we only read it
once when the server starts and never read it again, only when the server is
restarted. On the other hand, we write it frequently: we save the contents of
the word list object to the word list file after each modification.