Words have explicit indices
The words in the word list now have explicit indices. They are written as
"{index}" into the word list file. Words in the file without index are still
read and new indices are given to them.

The word list now stores the words in a dictionary instead of a list.