Bugfix: search page retains 'q' after action
Previously, if an empty query was initiated (which shows all word pairs in the
current dictionary), and an action (like deleting some word pairs) was
performed, the user was redirecting to a search page lacking the empty 'q'
parameter, which resulted in not showing any words.

Now we keep the empty 'q' parameter, so search page to which the user is
redirected will show all words in the dictionary, as expected.