Bugfix: label drop-down list uses exact match
Previously if the user selected e.g. the "new" label at the search page,
ExponWords showed word pairs with "new2" label as well. That is OK for search
queries (i.e. the "label:new" query), but not OK when the label is selected
in the drop-down list.

Now if the user selects a label from the drop-down list, only word pairs with
that exact tag will be shown.