Using ugettext_lazy
With plain ugettext, I had to use a trick (which is removed here), otherwise
the following error occurred:

    Let's suppose the user has an English browser. When he sets the language to
    "Hungarian", he will correctly see the message "Beállítások elmentve",
    which in Hungarian for "Settings saved". But when he afterwards sets the
    language to browser default, the message again will be "Beállítások
    elmentve", even though all other text appears on the page in English.

Now I use ugettext_lazy, and this way ExponWords behaves well even without the
trick.

A new argument 'source' was added to the import_* functions because so far
the translated text was inserted into the database, which is quite ugly.