Bugfix: "export textfile" removes newlines
Since the word forms are now edited in text areas (and not one liner input
fields), the users can insert newlines into them. This messes up the plain
text representation of the dictionary:

    ("Hi.\nI'm here.", "Szia.\nItt vagyok", ...)

becomes

    Hi.
    I'm here. -- Szia.
    Itt vagyok ...

on which the parser will fail. Therefore the new code replaces all newlines
with spaces as a workaround:

    Hi. I'm here. -- Szia. Itt vagyok ...

The proper solution to this problem will be using JSON instead of the plain
text format.