Newlines in word forms are displayed
So far we passed the words in raw form to the templates and JS code from
the Python code, and escaped only the explanation.

Now the word forms (word_in_lang1, word_in_lang2) are escaped and indented
just as the explanation. This effected only the practice page and the search
page.