Bugfix: dict pages appeared with wrong language
The "Edit dictionary" and "Create dictionary" pages had the text labels
sometimes appear using the wrong language.

The reason was that when the WDictForm class was defined, the translation of
the labels occurred according to the language settings at that time, and the
translations were never translated again, no matter which user with which
language setting loaded the form.