Practice page: load the words with AJAX again
Custom practice pages were added in commit 1624380455b105d0b6d47c8f1788, and in
the same commit view.py, practice_wdict.html and practice.js were modified so
that the list of words to practice was embedded in practice_wdict.html as a
JavaScript variable, instead of practice.js having to get it from the server
via an AJAX query calling get_words_to_practice_today.

Now this modification was reverted in the case when the practice page shows a
dictionary practice page (as opposed to a custom practice page). In the custom
case this modification is harder, because there is no
"get_words_to_custom_practice" query.