Make scrolling always work
This is a typical screen on mobile devices:

    +--------------------+
    | <menu>             |
    |                    |
    | 0(0) 0(0) 2 [More] |
    |                    |
    | +--------+         |
    | | ANSWER |         |
    | +--------+         |
    |                    |
    | Question ->        |
    |                    |
    |                    |
    |                    |
    +--------------------+

Before the previous commit (62b13c9925), the old default Android browser
allowed the screen to be scrolled e.g. to look like this:

    +--------------------+
    | +--------+         |
    | | ANSWER |         |
    | +--------+         |
    |                    |
    | Question ->        |
    |                    |
    |                    |
    |                    |
    |                    |
    |                    |
    |                    |
    |                    |
    +--------------------+

This scrolling ability is useful when the user uses the same hand to
hold the device (e.g. in the palm) and to click on the ANSWER button
(e.g. with the thumb), because in that case the user can comfortably
access only some areas of the screen.

After the previous commit (62b13c9925), the browsers no longer allow
that kind of scrolling (because they see that there is nothing on the
page to scroll to).

This commit puts a bunch of empty lines at the end of the HTML pages, so
that the browsers will allow this kind of scrolling.