Sort dictionaries in dropdown menus alphabetically
The "Search and operations" page shows a list of dictionaries in two
places:

1.  In a dropdown menu to select which dictionary to search in.

2.  In a dropdown menu to select which dictionary to move the selected
    words to.

Previously the dictionaries in these dropdown menus were in the same
order as sqlite returned them (which seems to be the order in which they
were created).

Now the dictionaries in these dropdown menus are ordered alphabetically.